+++ inst/tinytest/test_gettersSetters.R
@@ -16,7 +16,12 @@ x <- RPhosFate(
 source("parameters.R", local = TRUE) # nolint
 
 #### getLayer and [ extract operator ####
-layers <- list.files(cs_dir_ctl, "\\.tif$", full.names = TRUE, recursive = TRUE)
+layers <- list.files(
+  cs_dir_ctl,
+  "^\\D+\\.tif$",
+  full.names = TRUE,
+  recursive = TRUE
+)
 for (layer in layers) {
   expect_true(
     raster::all.equal(